[RDKEMW-4928]StorageManager Plugin updates for Singleton class
