+++ StorageManager/StorageManagerImplementation.h
@@ -49,29 +49,29 @@ namespace Plugin {
         };
 
     public:
-        typedef struct _StorageAppInfo
-        {
-            std::string path;    /* Path to the application's storage */
-            int32_t uid;         /* UID of the user who owns the storage */
-            int32_t gid;         /* GID of the group who owns the storage */
-            uint32_t quotaKB;    /* Quota size in kilobytes for the storage */
-            uint32_t usedKB;     /* Used space in kilobytes for the storage */
-            std::mutex storageLock; /* Mutex for thread safety */
-        } StorageAppInfo;
+        // typedef struct _StorageAppInfo
+        // {
+        //     std::string path;    /* Path to the application's storage */
+        //     int32_t uid;         /* UID of the user who owns the storage */
+        //     int32_t gid;         /* GID of the group who owns the storage */
+        //     uint32_t quotaKB;    /* Quota size in kilobytes for the storage */
+        //     uint32_t usedKB;     /* Used space in kilobytes for the storage */
+        //     std::mutex storageLock; /* Mutex for thread safety */
+        // } StorageAppInfo;
 
-        typedef struct _StorageSize
-        {
-            uint64_t blockSize = 0;
-            uint64_t usedBytes = 0;
-        } StorageSize;
+        // typedef struct _StorageSize
+        // {
+        //     uint64_t blockSize = 0;
+        //     uint64_t usedBytes = 0;
+        // } StorageSize;
 
-        enum StorageActionType
-        {
-            UNKNOWN = 0,
-            SET,
-            GET,
-            DELETE
-        };
+        // enum StorageActionType
+        // {
+        //     UNKNOWN = 0,
+        //     SET,
+        //     GET,
+        //     DELETE
+        // };
 
         StorageManagerImplementation();
         ~StorageManagerImplementation() override;
@@ -94,29 +94,18 @@ namespace Plugin {
         uint32_t Configure(PluginHost::IShell* service) override;
 
     private:
+         //bool isValidAppStorageDirectory(const std::string& dirName);
+         //Core::hresult createPersistentStoreRemoteStoreObject();
+         // void releasePersistentStoreRemoteStoreObject();
+        // Core::hresult populateAppInfoCacheFromStoragePath();
+        // Core::hresult appQuotaSizeProperty(StorageActionType actionType, const std::string& appId, uint32_t* quotaValue);
 
-        bool isValidAppStorageDirectory(const std::string& dirName);
-        Core::hresult createPersistentStoreRemoteStoreObject();
-        void releasePersistentStoreRemoteStoreObject();
-        Core::hresult appQuotaSizeProperty(StorageActionType actionType, const std::string& appId, uint32_t* quotaValue);
-        Core::hresult populateAppInfoCacheFromStoragePath();
-
-        bool createAppStorageInfoByAppID(const std::string& appId, StorageAppInfo &storageInfo);
-        bool retrieveAppStorageInfoByAppID(const string &appId, StorageAppInfo &storageInfo);
-        bool removeAppStorageInfoByAppID(const string &appId);
-        bool hasEnoughStorageFreeSpace(const std::string& baseDir, uint32_t requiredSpaceKB);
-        uint64_t getDirectorySizeInBytes(const std::string &path);
-        static int getSize(const char *path, const struct stat *statPtr, int currentFlag, struct FTW *internalFtwUsage);
-        Core::hresult deleteDirectoryEntries(const string& appId, string& errorReason);
-        bool lockAppStorageInfo(const std::string& appId, std::unique_lock<std::mutex>& appLock);
-
-    private:
-        mutable std::mutex mStorageManagerImplLock;
-        static std::mutex mStorageSizeLock;
-        std::map<std::string, StorageAppInfo> mStorageAppInfo;  /* Map storing app storage info for each appId */
+       // mutable std::mutex mStorageManagerImplLock;
+       // static std::mutex mStorageSizeLock;
+        //std::map<std::string, StorageAppInfo> mStorageAppInfo;  /* Map storing app storage info for each appId */
         Config _config;
         PluginHost::IShell* mCurrentservice;
-        Exchange::IStore2* mPersistentStoreRemoteStoreObject;
+        // Exchange::IStore2* mPersistentStoreRemoteStoreObject;
         std::string mBaseStoragePath;                           /* Base path for the app storage */
     };
 } /* namespace Plugin */
